+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 
+ * Support for *importing* VEDA data files.
  * Tests of ``__main__``.
  * Added tests and documentation of `init`, `init_cols`, `init_icols`, `auto_init` and corresponding post keyword arguments of wrapper functions.